Add files via upload
A Zero Indexed Array A consisting of N integers is given. A pair of integers (P,Q) such that 0<=P<=Q

class Solution { 
public int solution(int[] A); 
}
that, given an Array A consisting of N integers, returns the size of the largest bi-valued slice in A.

For example : A[0] = 5, A[1] = 4, A[2] = 4, A[3] = 5, A[4] = 0, A[5] = 12 the function should return 4, because (0,3) is a slice containing only 4 and 5. Note that a bi-valued slice can also contain just one value. Slice(3,5) is not bi-valued since it consists of three different numbers.